// server/src/generated/models.ts
/* eslint-disable */
// tslint:disable
/**
 * Secure Chat API
 * This is a secure chat API.
You can find out more at [secure-chat-app](https://github.com/Sriram-52/secure-chat).

 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

export namespace Api {
	export interface Message {
		/**
		 * <p>ID of the message</p>
		 * @type {string}
		 * @memberof Message
		 */
		id?: string;
		/**
		 * <p>userid of the sender</p>
		 * @type {string}
		 * @memberof Message
		 */
		sender: string;
		/**
		 * <p>userid of the receiver</p>
		 * @type {string}
		 * @memberof Message
		 */
		receiver: string;
		/**
		 * <p>Message to send</p>
		 * @type {string}
		 * @memberof Message
		 */
		message: string;
	}

	export interface User {
		/**
		 * <p>ID of the user</p>
		 * @type {string}
		 * @memberof User
		 */
		id?: string;
		/**
		 * <p>Username of the user</p>
		 * @type {string}
		 * @memberof User
		 */
		username: string;
		/**
		 * <p>Email of the user</p>
		 * @type {string}
		 * @memberof User
		 */
		email: string;
		/**
		 * <p>Password of the user</p>
		 * @type {string}
		 * @memberof User
		 */
		password?: string;
	}

}